Make the query work, with a caveat
DETAILS:
Redesigned as union of two queries. Both queries work.

HOWEVER, the full MESH query for definitions via
meshv:scopNote (line commented out) returns 0 results.
The reason for this appears to be that there are no
matches to the meshv:scopeNote predicate for DOID-MESH xrefs.
In its current form it returns a long list of predicate/object
results but none appear to be definitions.