Initiate notebook: DO & ICD-O term fuzzy matching
Designed To identify potential mappings between DO and ICD-O at
the request of ClinGen, this includes initial analysis of label
length and formatting.

[ON PAUSE] INDRA has a more powerful matching algorithm that
might be able to handle the ICD-O data more intelligently than
simple approximate string matching. We will give that approach
an attempt first and return to this one only if necessary.