Fix reticulate-python install
PROBLEM:
install_python() fails to complete installation the first time
it is run. This error is documented in reticulate's github
issues: https://github.com/rstudio/reticulate/issues/944

FIX:
Repeat call to install_python() if it does not succeed on the
first try. It has been successful on the second attempt and two
attempts seem to be necessary currently as the first installs
some support software that's required for success.