Change how Babel "fuzzy" translations are handled
Updated version of Babel is much more likely to create "fuzzy"
translations, which caused frequent errors.

This downgrades 'fuzzy' translations to warnings and, instead of
treating them as text in need of updating, sets the fuzzy
translations aside to promote a simpler workflow with Google
translating from scratch (fuzzy translations are often wrong and
more confusing to review).